requirements.txt
numpy>=1.24
scipy>=1.10
pandas>=2.1
scikit-learn>=1.3
biopython>=1.81
goatools>=1.3.0
joblib>=1.3
tqdm>=4.66
